+++ src/main/java/org/opentripplanner/graph_builder/GraphBuilder.java
@@ -93,7 +93,9 @@ public class GraphBuilder implements Runnable {
     public static GraphBuilder create(
             BuildConfig config,
             GraphBuilderDataSources dataSources,
-            Graph baseGraph
+            Graph baseGraph,
+            boolean loadStreetGraph,
+            boolean saveStreetGraph
     ) {
 
         boolean hasOsm  = dataSources.has(OSM);
@@ -132,13 +134,6 @@ public class GraphBuilder implements Runnable {
 
                 GtfsBundle gtfsBundle = new GtfsBundle((CompositeDataSource)gtfsData);
 
-                // TODO OTP2 - In OTP2 we have deleted the transfer edges from the street graph.
-                //           - The new transfer generation do not take this config param into
-                //           - account any more. This needs some investigation and probably
-                //           - a fix, but we are unsure if this is used any more. The Pathways.txt
-                //           - and osm import replaces this functionality.
-                gtfsBundle.setTransfersTxtDefinesStationPaths(config.useTransfersTxt);
-
                 if (config.parentStopLinking) {
                     gtfsBundle.linkStopsToParentStations = true;
                 }
@@ -172,7 +167,7 @@ public class GraphBuilder implements Runnable {
         // Prune graph connectivity islands after transit stop linking, so that pruning can take into account
         // existence of stops in islands. If an island has a stop, it actually may be a real island and should
         // not be removed quite as easily
-        if ( hasOsm ) {
+        if ((hasOsm && !saveStreetGraph) || loadStreetGraph) {
             PruneNoThruIslands pruneNoThruIslands = new PruneNoThruIslands(streetLinkerModule);
             pruneNoThruIslands.setPruningThresholdIslandWithoutStops(config.pruningThresholdIslandWithoutStops);
             pruneNoThruIslands.setPruningThresholdIslandWithStops(config.pruningThresholdIslandWithStops);
@@ -220,11 +215,10 @@ public class GraphBuilder implements Runnable {
             if (OTPFeature.FlexRouting.isOn()) {
                 graphBuilder.addModule(new FlexLocationsToStreetEdgesMapper());
             }
-            // The stops can be linked to each other once they are already linked to the street network.
-            if ( ! config.useTransfersTxt) {
-                // This module will use streets or straight line distance depending on whether OSM data is found in the graph.
-                graphBuilder.addModule(new DirectTransferGenerator(config.maxTransferDurationSeconds, config.transferRequests));
-            }
+
+            // This module will use streets or straight line distance depending on whether OSM data is found in the graph.
+            graphBuilder.addModule(new DirectTransferGenerator(config.maxTransferDurationSeconds, config.transferRequests));
+
             // Analyze routing between stops to generate report
             if (OTPFeature.TransferAnalyzer.isOn()) {
                 graphBuilder.addModule(new DirectTransferAnalyzer(