+++ src/test/java/org/opentripplanner/gtfs/mapping/PathwayNodeMapperTest.java
@@ -74,8 +74,8 @@ public class PathwayNodeMapperTest {
     assertEquals("A:N1", result.getId().toString());
     assertEquals(CODE, result.getCode());
     assertEquals(DESC, result.getDescription());
-    assertEquals(LAT, result.getLat(), 0.0001d);
-    assertEquals(LON, result.getLon(), 0.0001d);
+    assertEquals(LAT, result.getCoordinate().latitude(), 0.0001d);
+    assertEquals(LON, result.getCoordinate().longitude(), 0.0001d);
     assertEquals(NAME, result.getName());
     assertEquals(WheelChairBoarding.POSSIBLE, result.getWheelchairBoarding());
   }
@@ -105,7 +105,7 @@ public class PathwayNodeMapperTest {
 
     org.opentripplanner.model.PathwayNode result = subject.map(input);
 
-    result.getLat();
+    result.getCoordinate().latitude();
   }
 
   /** Mapping the same object twice, should return the the same instance. */